Move GameController and MapController from org.rpgrunner to org.rpgrunner.controller
