+++ src/test/java/org/rpgrunner/test/mock/controller/MapControllerSpy.java
@@ -1,6 +1,6 @@
-package org.rpgrunner.test.mock;
+package org.rpgrunner.test.mock.controller;
 
-import org.rpgrunner.MapController;
+import org.rpgrunner.controller.MapController;
 import org.rpgrunner.character.CharacterElement;
 import org.rpgrunner.event.action.Action;
 import org.rpgrunner.helper.MapHelper;
