+++ src/test/java/org/rpgrunner/test/mock/event/factory/ActionAbstractFactorySpy.java
@@ -5,10 +5,10 @@ import java.io.InputStream;
 import java.util.Enumeration;
 import java.util.Vector;
 
-import org.rpgrunner.GameController;
+import org.rpgrunner.controller.GameController;
 import org.rpgrunner.event.action.Action;
 import org.rpgrunner.event.factory.ActionAbstractFactory;
-import org.rpgrunner.test.mock.GameControllerSpy;
+import org.rpgrunner.test.mock.controller.GameControllerSpy;
 import org.rpgrunner.test.mock.event.action.ActionSpy;
 
 public class ActionAbstractFactorySpy extends ActionAbstractFactory {
