+++ src/test/java/org/rpgrunner/test/mock/helper/MapHelperSpy.java
@@ -3,7 +3,7 @@ package org.rpgrunner.test.mock.helper;
 import org.rpgrunner.character.GameCharacter;
 import org.rpgrunner.helper.MapHelper;
 import org.rpgrunner.map.Map;
-import org.rpgrunner.test.mock.GameControllerSpy;
+import org.rpgrunner.test.mock.controller.GameControllerSpy;
 
 public class MapHelperSpy extends MapHelper {
     private boolean canMoveValue;
